[scanner] refactor: split pkg/agent/kube/client.go into focused modules
